refactor: 471 - rename entity. Move Main & General. Fix imports
