+++ src/entities/course/index.ts
@@ -1,0 +1,1 @@
+export { CourseCard } from './ui/course-card/course-card';
