+++ src/entities/course/types.ts
