+++ src/entities/course/ui/course-card/course-card.scss
