--- src/entities/courses/index.ts
@@ -1,3 +1,0 @@
-export { CourseCard } from './ui/course-card/course-card';
-export { Main as CoursesMain } from './ui/main/main';
-export { General } from './ui/general/general';
