+++ src/pages/courses.tsx
@@ -1,8 +1,9 @@
 import { FC } from 'react';
-import { CoursesMain, General } from '@/entities/courses';
 import { useTitle } from '@/shared/hooks/use-title';
 import { Breadcrumbs } from '@/widgets/breadcrumbs';
 import { RSCourses } from '@/widgets/courses';
+import { General } from '@/widgets/general';
+import { CoursesMain } from '@/widgets/main';
 import { StudyPath } from '@/widgets/study-path';
 
 export const Courses: FC = () => {
