+++ src/shared/helpers/is-course.ts
