+++ src/widgets/courses/ui/courses.tsx
@@ -1,8 +1,8 @@
 import { COURSE_STALE_AFTER_DAYS } from '@/app/const';
 import { Course } from '@/app/types';
-import { CourseCard } from '@/entities/courses';
-import { isCourse } from '@/entities/courses/helpers/is-course';
+import { CourseCard } from '@/entities/course';
 import { getActualData } from '@/shared/helpers/getActualData';
+import { isCourse } from '@/shared/helpers/is-course';
 import { useDataByName } from '@/shared/hooks/use-data-by-name';
 import { WidgetTitle } from '@/shared/ui/widget-title';
