+++ src/widgets/general/general.scss
