+++ src/widgets/general/general.tsx
