+++ src/widgets/main/index.ts
@@ -1,1 +1,2 @@
+export { Main as CoursesMain } from './ui/main/main';
 export { Main } from './ui/main';
