+++ src/widgets/main/ui/main/main.scss
